lweuiu

+++ src/utils.ts
@@ -104,6 +104,8 @@ export const defaultMessages = [
 ];
 
 export async function startNewUserProcess(error: any, clientId: string) {
+
+
   if (error.errorMessage == 'CONNECTION_NOT_INITED' || error.errorMessage == 'AUTH_KEY_DUPLICATED') {
     await fetchWithTimeout(`${ppplbot()}&text=@${(process.env.clientId).toUpperCase()}-${clientId}: AUTH KEY DUPLICATED : Deleting Archived Client`);
     console.log("AUTH KEY DUPLICATED : Deleting Archived Client")
@@ -111,28 +113,31 @@ export async function startNewUserProcess(error: any, clientId: string) {
     process.exit(1);
   }
   if (error.errorMessage === "USER_DEACTIVATED_BAN" || error.errorMessage === "USER_DEACTIVATED") {
-    console.log(`${(process.env.clientId).toUpperCase()}-${clientId} USER_DEACTIVATED : Exitiing`)
-    await fetchWithTimeout(`${ppplbot()}&text=@${(process.env.clientId).toUpperCase()}-${clientId}: USER_DEACTIVATED : Exitiing`);
     const db = UserDataDtoCrud.getInstance();
-    const today = (new Date(Date.now())).toISOString().split('T')[0];
-    const query = { availableDate: { $lte: today }, channels: { $gt: 200 } }
-    const newPromoteClient = await db.findPromoteClient(query)
-    if (newPromoteClient) {
-      console.log(clientId, " - NEw Promote Client: ", newPromoteClient)
-      await db.updateClient(
-        {
-          clientId: clientId
-        },
-        {
-          promoteMobile: newPromoteClient.mobile
-        }
-      )
-      const clientDetails = getClient(clientId)
-      await db.deletePromoteClient({ mobile: clientDetails.mobile })
-    } else {
-      console.log(`${(process.env.clientId).toUpperCase()}-${clientId} new client does not exist`)
+    const clientDetails = getClient(clientId);
+    const actualClient = await db.getClient({ clientId });
+    if (actualClient.promoteMobile == clientDetails.mobile) {
+      console.log(`${(process.env.clientId).toUpperCase()}-${clientId} USER_DEACTIVATED : Exitiing`)
+      await fetchWithTimeout(`${ppplbot()}&text=@${(process.env.clientId).toUpperCase()}-${clientId}: USER_DEACTIVATED : Exitiing`);
+      const today = (new Date(Date.now())).toISOString().split('T')[0];
+      const query = { availableDate: { $lte: today }, channels: { $gt: 200 } }
+      const newPromoteClient = await db.findPromoteClient(query)
+      if (newPromoteClient) {
+        console.log(clientId, " - NEw Promote Client: ", newPromoteClient)
+        await db.updateClient(
+          {
+            clientId: clientId
+          },
+          {
+            promoteMobile: newPromoteClient.mobile
+          }
+        )
+        await db.deletePromoteClient({ mobile: clientDetails.mobile })
+      } else {
+        console.log(`${(process.env.clientId).toUpperCase()}-${clientId} new client does not exist`)
+      }
+      // process.exit(1)
     }
-    // process.exit(1)
   }
 }
